requirements.txt
torch>=2.0
numpy
Pillow
matplotlib
opencv-python
imageio
tqdm
